Allow a user to supply a sorted list of channels to be mapped
In the UK (at least) there are many channels broadcast, many of
which a user won't want to see, and they're often in a random or
location-dependent order. This edit allows the user to supply a
sorted list of TVHeadend channel numbers, which will then be used
to sort and filter the channel listing given to Plex.
In an ideal world, Plex's DVR feature would have its own sensible
filter and sorting, but it hasn't happened yet. At least in this
Python file it won't be lost if you rebuild Plex.
Example use:
        'channelMapping': [
            115, # bbc one hd
            802, # bbc two hd
            803, # ITV HD
            974 # channel 4
         ]